refactor: modified build
