+++ .flowconfig
@@ -3,9 +3,9 @@
 [include]
 
 [libs]
+./packages/ast/src/types/
 ./types/
 ./packages/wasm-parser/src/types/decoder.js
-./packages/ast/src/types.js
 
 [lints]
